refactor(Questionnaire): Split hook
